# server/internal/proto/api/base/setting.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/base/setting.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "SettingService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/setting/{key}": {
      "get": {
        "operationId": "SettingService_GetSetting",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/SettingGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "SettingService"
        ]
      }
    },
    "/api/settings": {
      "post": {
        "operationId": "SettingService_UpdateSetting",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseSettingUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/baseSettingUpdateRequest"
            }
          }
        ],
        "tags": [
          "SettingService"
        ]
      }
    },
    "/api/settings/reset": {
      "post": {
        "operationId": "SettingService_ResetSetting",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/SettingResetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SettingService"
        ]
      }
    }
  },
  "definitions": {
    "SettingGetResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/baseSetting"
        }
      }
    },
    "SettingResetResponse": {
      "type": "object"
    },
    "baseSetting": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "title": "@gotags: gorm:\"primary_key;auto_increment\""
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "title": "@gotags: gorm:\"serializer:protobuf_timestamp;type:datetime\""
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "title": "@gotags: gorm:\"serializer:protobuf_timestamp;type:datetime\""
        },
        "key": {
          "type": "string",
          "title": "@gotags: gorm:\"size:128;not null;unique\""
        },
        "value": {
          "type": "string"
        }
      }
    },
    "baseSettingUpdateRequest": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "baseSettingUpdateResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/baseSetting"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}